+++ dst/app/routes/ownershipInfos.js
@@ -58,20 +58,25 @@ ownershipInfosRouter.post('/saveByIdentifier', permitScopes_1.default([]), ...[
         const ownershipInfo = yield ownershipInfoRepo.saveByIdentifier(Object.assign(Object.assign({ id: '', identifier: req.body.identifier, ownedBy: req.body.ownedBy, ownedFrom: req.body.ownedFrom, project: { typeOf: chevre.factory.organizationType.Project, id: req.project.id }, typeOf: 'OwnershipInfo', typeOfGood: req.body.typeOfGood }, (req.body.ownedThrough instanceof Date) ? { ownedThrough: req.body.ownedThrough } : undefined), (req.body.acquiredFrom !== undefined && req.body.acquiredFrom !== null)
             ? { acquiredFrom: req.body.acquiredFrom }
             : undefined));
-        // 不要な所有権を削除
-        yield ownershipInfoRepo.ownershipInfoModel.deleteMany({
-            'project.id': { $eq: req.project.id },
-            // 1年以上前に所有したもの
-            ownedFrom: {
-                $lt: moment(now)
-                    // tslint:disable-next-line:no-magic-numbers
-                    .add(-12, 'months')
-                    .toDate()
-            },
-            // 所有期限切れのもの(ownedThroughの存在しないものは削除してはいけない)
-            ownedThrough: { $exists: true, $lt: now }
-        })
-            .exec();
+        try {
+            // 不要な所有権を削除
+            yield ownershipInfoRepo.ownershipInfoModel.deleteMany({
+                'project.id': { $eq: req.project.id },
+                // 1年以上前に所有したもの
+                ownedFrom: {
+                    $lt: moment(now)
+                        // tslint:disable-next-line:no-magic-numbers
+                        .add(-12, 'months')
+                        .toDate()
+                },
+                // 所有期限切れのもの(ownedThroughの存在しないものは削除してはいけない)
+                ownedThrough: { $exists: true, $lt: now }
+            })
+                .exec();
+        }
+        catch (error) {
+            console.error('ownershipInfoRepo.ownershipInfoModel.deleteMany throws', error);
+        }
         res.status(http_status_1.CREATED)
             .json(ownershipInfo);
     }