+++ src/app/routes/ownershipInfos.ts
@@ -68,20 +68,24 @@ ownershipInfosRouter.post(
                     : undefined
             });
 
-            // 不要な所有権を削除
-            await ownershipInfoRepo.ownershipInfoModel.deleteMany({
-                'project.id': { $eq: req.project.id },
-                // 1年以上前に所有したもの
-                ownedFrom: {
-                    $lt: moment(now)
-                        // tslint:disable-next-line:no-magic-numbers
-                        .add(-12, 'months')
-                        .toDate()
-                },
-                // 所有期限切れのもの(ownedThroughの存在しないものは削除してはいけない)
-                ownedThrough: { $exists: true, $lt: now }
-            })
-                .exec();
+            try {
+                // 不要な所有権を削除
+                await ownershipInfoRepo.ownershipInfoModel.deleteMany({
+                    'project.id': { $eq: req.project.id },
+                    // 1年以上前に所有したもの
+                    ownedFrom: {
+                        $lt: moment(now)
+                            // tslint:disable-next-line:no-magic-numbers
+                            .add(-12, 'months')
+                            .toDate()
+                    },
+                    // 所有期限切れのもの(ownedThroughの存在しないものは削除してはいけない)
+                    ownedThrough: { $exists: true, $lt: now }
+                })
+                    .exec();
+            } catch (error) {
+                console.error('ownershipInfoRepo.ownershipInfoModel.deleteMany throws', error);
+            }
 
             res.status(CREATED)
                 .json(ownershipInfo);