+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - IAMロール検索条件拡張
 - 予約にissuedThroughを追加
 - 予約のreservationForを最適化
+- 所有権作成時に不要な所有権を削除する処理を追加
 
 ### Deprecated