+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 予約にissuedThroughを追加
 - 予約のreservationForを最適化
 - 所有権作成時に不要な所有権を削除する処理を追加
+- 注文作成時に不要な注文を削除する処理を追加
+- プロダクトのserviceTypeをサービスアウトプットに連携
 
 ### Deprecated